stream serialize
